+++ components/admin/data/widgets/form/steps/component.jsx
@@ -10,8 +10,15 @@ import Field from 'components/form/Field';
 import Select from 'components/form/SelectInput';
 import Checkbox from 'components/form/Checkbox';
 
+// constants
+import {
+  WIDGET_EDITOR_COLOUR_SCHEMES,
+} from 'constants/widget-editor';
+
 // Utils
-import DefaultTheme from 'utils/widgets/theme';
+import {
+  getDefaultTheme,
+} from 'utils/widget-editor';
 
 class AdminWidgetForm extends Component {
   constructor(props) {
@@ -169,8 +176,9 @@ class AdminWidgetForm extends Component {
             {...(id && { widgetId: id })}
             application="rw"
             onSave={onSave}
-            theme={DefaultTheme}
+            theme={getDefaultTheme}
             adapter={RWAdapter}
+            schemes={WIDGET_EDITOR_COLOUR_SCHEMES}
             authenticated
             compact={false}
           />