+++ components/app/myrw/widgets/tabs/edit/component.jsx
@@ -14,8 +14,15 @@ import {
   updateWidgetMetadata,
 } from 'services/widget';
 
+// constants
+import {
+  WIDGET_EDITOR_COLOUR_SCHEMES,
+} from 'constants/widget-editor';
+
 // utils
-import DefaultTheme from 'utils/widgets/theme';
+import {
+  getDefaultTheme,
+} from 'utils/widget-editor';
 
 class MyRWWidgetEditTab extends React.Component {
   constructor(props) {
@@ -112,8 +119,9 @@ class MyRWWidgetEditTab extends React.Component {
             datasetId={widget.dataset}
             widgetId={widget.id}
             onSave={this.onSaveWidget}
-            theme={DefaultTheme}
+            theme={getDefaultTheme}
             adapter={RWAdapter}
+            schemes={WIDGET_EDITOR_COLOUR_SCHEMES}
           />
         )}
       </div>